* change the result style of human_preference_annotation_mapper from "result" to "chosen/rejected"
